@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "BuildingMOTIF"
-version = "0.3.0"
+version = "0.4.0"
 description = "Building Metadata OnTology Interoperability Framework"
 license = "LICENSE"
 authors = [
@@ -19,26 +19,28 @@ documentation = "https://buildingmotif.readthedocs.io/"
 buildingmotif = 'buildingmotif.bin.cli:app'
 
 [tool.poetry.dependencies]
-python = ">=3.8.1, <3.12"
+python = ">=3.10, <3.13"
 rdflib = ">=7.0"
 SQLAlchemy = {extras = ["mypy"], version = "^1.4.44"}
 pyaml = "^21.10.1"
-networkx = "^2.7.1"
+networkx = "^3.4.2"
 types-PyYAML = "^6.0.4"
 nbmake = "^1.3.0"
 rdflib-sqlalchemy = "^0.5.3"
-pyshacl = "^0.26"
+pyshacl = "^0.30"
 alembic = "^1.8.0"
 Flask = "^2.1.2"
 Flask-API = "^3.0.post1"
 rfc3987 = "^1.3.8"
 setuptools = "^73.0.1"
 psycopg2 = {version="^2.9.5", optional=true}
-pygit2 = "~1.11.1"
+pygit2 = "~1.17"
 jsonschema = "^4.21.1"
-brick-tq-shacl = {optional = true, version="0.3.3a3"}
+brick-tq-shacl = {optional = true, version="0.3.4a2"}
 werkzeug="^2.3.7"
 types-jsonschema = "^4.21.0.20240311"
+matplotlib = "^3.9.2"
+pandas = "^2.2.3"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.3.0"
@@ -53,12 +55,13 @@ jupyter = "^1.0.0"
 ipykernel = "^6.13.0"
 rise = "^5.7.1"
 jupyter-book = "^0.15.1"
-flake8 = "^5.0.0"
+flake8 = "^7.0.0"
 BAC0 = "^22.9.21"
 netifaces = "^0.11.0"
 pytz = "^2022.7.1"
 openpyxl = "^3.0.10"
 pytest = "^8.0.2"
+tomlkit = "^0.13.2"
 
 [tool.poetry.extras]
 all = ["BAC0", "openpyxl", "netifaces", "pytz", "psycopg2"]
