@@ -1,12 +1,32 @@
-from typing import Dict, List
+from typing import Dict, List, Optional
 
-from sqlalchemy import Column, ForeignKey, Integer, String, Text, UniqueConstraint
-from sqlalchemy.orm import Mapped, declarative_base, relationship
+from sqlalchemy import (
+    Column,
+    ForeignKey,
+    Integer,
+    String,
+    Text,
+    UniqueConstraint,
+    event,
+    select,
+)
+from sqlalchemy.engine import Engine
+from sqlalchemy.exc import NoResultFound
+from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy.orm import Mapped, Session, declarative_base, relationship
 
 # from sqlalchemy.dialects.postgresql import JSON
 from buildingmotif.database.utils import JSONType
 
 Base = declarative_base()
+
+
+# https://docs.sqlalchemy.org/en/14/dialects/sqlite.html#foreign-key-support
+@event.listens_for(Engine, "connect")
+def set_sqlite_pragma(dbapi_connection, connection_record):
+    cursor = dbapi_connection.cursor()
+    cursor.execute("PRAGMA foreign_keys=ON")
+    cursor.close()
 
 
 class DBModel(Base):
@@ -18,12 +38,13 @@ class DBModel(Base):
     description: Mapped[str] = Column(Text(), default="", nullable=False)
     graph_id: Mapped[str] = Column(String())
     manifest_id: Mapped[int] = Column(
-        Integer, ForeignKey("shape_collection.id"), nullable=False
+        Integer, ForeignKey("shape_collection.id", ondelete="CASCADE"), nullable=False
     )
     manifest: "DBShapeCollection" = relationship(
         "DBShapeCollection",
         uselist=False,
-        cascade="all,delete",
+        cascade="all",
+        passive_deletes=True,
     )
 
 
@@ -44,38 +65,19 @@ class DBLibrary(Base):
     id: Mapped[int] = Column(Integer, primary_key=True)
     name: Mapped[str] = Column(String(), nullable=False, unique=True)
 
+    # do not use passive_deletes here because we want to handle the deletion of templates
     templates: Mapped[List["DBTemplate"]] = relationship(
-        "DBTemplate", back_populates="library", cascade="all,delete"
+        "DBTemplate", back_populates="library", cascade="all"
     )
 
     shape_collection_id = Column(
-        Integer, ForeignKey("shape_collection.id"), nullable=False
+        Integer, ForeignKey("shape_collection.id", ondelete="CASCADE"), nullable=False
     )
     shape_collection: DBShapeCollection = relationship(
         "DBShapeCollection",
         uselist=False,
-        cascade="all,delete",
-    )
-
-
-class DepsAssociation(Base):
-    """Many-to-many relationship between dependant templates."""
-
-    __tablename__ = "deps_association_table"
-
-    id: Mapped[int] = Column(Integer, primary_key=True)
-    dependant_id: Mapped[int] = Column(ForeignKey("template.id"))
-    dependee_id: Mapped[int] = Column(ForeignKey("template.id"))
-    # args are a mapping of dependee args to dependant args
-    args: Mapped[Dict[str, str]] = Column(JSONType)  # type: ignore
-
-    __table_args__ = (
-        UniqueConstraint(
-            "dependant_id",
-            "dependee_id",
-            "args",
-            name="deps_association_unique_constraint",
-        ),
+        cascade="all",
+        passive_deletes=True,
     )
 
 
@@ -89,21 +91,13 @@ class DBTemplate(Base):
     body_id: Mapped[str] = Column(String())
     optional_args: Mapped[List[str]] = Column(JSONType)  # type: ignore
 
-    library_id: Mapped[int] = Column(Integer, ForeignKey("library.id"), nullable=False)
-    library: Mapped[DBLibrary] = relationship("DBLibrary", back_populates="templates")
-    dependencies: Mapped[List["DBTemplate"]] = relationship(
-        "DBTemplate",
-        secondary="deps_association_table",
-        primaryjoin=id == DepsAssociation.dependant_id,
-        secondaryjoin=id == DepsAssociation.dependee_id,
-        back_populates="dependants",
+    library_id: Mapped[int] = Column(
+        Integer, ForeignKey("library.id", ondelete="CASCADE"), nullable=False
     )
-    dependants: Mapped[List["DBTemplate"]] = relationship(
-        "DBTemplate",
-        secondary="deps_association_table",
-        primaryjoin=id == DepsAssociation.dependee_id,
-        secondaryjoin=id == DepsAssociation.dependant_id,
-        back_populates="dependencies",
+    library: Mapped[DBLibrary] = relationship("DBLibrary", back_populates="templates")
+
+    dependencies: Mapped["DBTemplateDependency"] = relationship(
+        "DBTemplateDependency", back_populates="template", cascade="all,delete-orphan"
     )
 
     __table_args__ = (
@@ -111,5 +105,60 @@ class DBTemplate(Base):
             "name",
             "library_id",
             name="name_library_unique_constraint",
+        ),
+    )
+
+
+class DBTemplateDependency(Base):
+    __tablename__ = "template_dependency"
+    id: Mapped[int] = Column(Integer, primary_key=True)
+
+    template_id: Mapped[int] = Column(
+        Integer, ForeignKey("template.id", ondelete="CASCADE"), nullable=False
+    )
+    template: Mapped[DBTemplate] = relationship(
+        DBTemplate, back_populates="dependencies"
+    )
+
+    dependency_library_name: Mapped[str] = Column(String, nullable=False)
+    dependency_template_name: Mapped[str] = Column(String, nullable=False)
+
+    # args are a mapping of dependee args to dependant args
+    args: Mapped[Dict[str, str]] = Column(JSONType)  # type: ignore
+
+    @hybrid_property
+    def dependency_template(self) -> Optional[DBTemplate]:
+        session = Session.object_session(self)
+        statement = (
+            select(DBTemplate)
+            .join(DBTemplate.library)
+            .where(
+                DBTemplate.name == self.dependency_template_name,
+                DBLibrary.name == self.dependency_library_name,
+            )
+        )
+        try:
+            return session.scalars(statement).one()
+        except NoResultFound:
+            return None
+
+    @dependency_template.expression
+    def _dependency_tempalate(self):
+        return (
+            select(DBTemplate)
+            .join(DBTemplate.library)
+            .where(
+                DBTemplate.name == self.dependency_template_name,
+                DBLibrary.name == self.dependency_library_name,
+            )
+        )
+
+    __table_args__ = (
+        UniqueConstraint(
+            "template_id",
+            "dependency_library_name",
+            "dependency_template_name",
+            "args",
+            name="template_dependency_unique_constraint",
         ),
     )
