@@ -3,9 +3,13 @@ from flask import Blueprint, current_app, jsonify, request
 from flask_api import status
 from rdflib import Graph, URIRef
 from rdflib.plugins.parsers.notation3 import BadSyntax
-from sqlalchemy.orm.exc import NoResultFound
 
 from buildingmotif.api.serializers.model import serialize
+from buildingmotif.database.errors import (
+    LibraryNotFound,
+    ModelNotFound,
+    ShapeCollectionNotFound,
+)
 from buildingmotif.dataclasses import Library, Model, ShapeCollection
 
 blueprint = Blueprint("models", __name__)
@@ -34,8 +38,8 @@ def get_model(models_id: int) -> flask.Response:
     """
     try:
         model = current_app.building_motif.table_connection.get_db_model(models_id)
-    except NoResultFound:
-        return {"message": f"No model with id {models_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {models_id}"}, status.HTTP_404_NOT_FOUND
 
     return jsonify(serialize(model)), status.HTTP_200_OK
 
@@ -51,8 +55,8 @@ def get_model_graph(models_id: int) -> Graph:
     """
     try:
         model = Model.load(models_id)
-    except NoResultFound:
-        return {"message": f"No model with id {models_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {models_id}"}, status.HTTP_404_NOT_FOUND
 
     g = Graph() + model.graph
 
@@ -70,8 +74,8 @@ def get_target_nodes(models_id: int) -> Graph:
     """
     try:
         model = Model.load(models_id)
-    except NoResultFound:
-        return {"message": f"No model with id {models_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {models_id}"}, status.HTTP_404_NOT_FOUND
 
     result = model.graph.query(
         """
@@ -132,8 +136,8 @@ def update_model_graph(models_id: int) -> flask.Response:
     """
     try:
         model = Model.load(models_id)
-    except NoResultFound:
-        return {"message": f"No model with id {models_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {models_id}"}, status.HTTP_404_NOT_FOUND
 
     if request.content_type != "application/xml":
         return {
@@ -160,12 +164,16 @@ def validate_model(models_id: int) -> flask.Response:
     # get model
     try:
         model = Model.load(models_id)
-    except NoResultFound:
-        return {"message": f"No model with id {models_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {models_id}"}, status.HTTP_404_NOT_FOUND
 
+    # we will read the shape collections from the input
     shape_collections = []
 
-    # no body provided -- default to model manifest and default SHACL engine
+    # get shacl_engine from the query params, defaults to the current engine
+    shacl_engine = request.args.get("shacl_engine", None)
+
+    # no body provided -- default to model manifest
     if request.content_length is None:
         shape_collections = [model.get_manifest()]
     else:
@@ -182,23 +190,23 @@ def validate_model(models_id: int) -> flask.Response:
 
         if body is not None and not isinstance(body, dict):
             return {"message": "body is not dict"}, status.HTTP_400_BAD_REQUEST
-        shape_collections = []
         body = body if body is not None else {}
         nonexistent_libraries = []
         for library_id in body.get("library_ids", []):
             try:
                 shape_collection = Library.load(library_id).get_shape_collection()
                 shape_collections.append(shape_collection)
-            except NoResultFound:
+            except LibraryNotFound:
                 nonexistent_libraries.append(library_id)
         if len(nonexistent_libraries) > 0:
             return {
                 "message": f"Libraries with ids {nonexistent_libraries} do not exist"
             }, status.HTTP_400_BAD_REQUEST
 
-    # if shape_collections is empty, model.validate will default
-    # to the model's manifest
-    vaildation_context = model.validate(shape_collections)
+    # if shape_collections is empty, model.validate will default to the model's manifest
+    vaildation_context = model.validate(
+        shape_collections, error_on_missing_imports=False, shacl_engine=shacl_engine
+    )
 
     return {
         "message": vaildation_context.report_string,
@@ -215,8 +223,8 @@ def validate_shape(models_id: int) -> flask.Response:
     # get model
     try:
         model = Model.load(models_id)
-    except NoResultFound:
-        return {"message": f"No model with id {models_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {models_id}"}, status.HTTP_404_NOT_FOUND
 
     # get body
     if request.content_type != "application/json":
@@ -239,7 +247,7 @@ def validate_shape(models_id: int) -> flask.Response:
         try:
             shape_collection = ShapeCollection.load(shape_collection_id)
             shape_collections.append(shape_collection)
-        except NoResultFound:
+        except ShapeCollectionNotFound:
             nonexistent_shape_collections.append(shape_collection_id)
     if len(nonexistent_shape_collections) > 0:
         return {
@@ -255,8 +263,8 @@ def validate_shape(models_id: int) -> flask.Response:
     target_class = URIRef(body.get("target_class"))
 
     # test
-    conformance = model.test_model_against_shapes(
-        shape_collections=shape_collections,
+    compiled = model.compile(shape_collections)
+    conformance = compiled.validate_model_against_shapes(
         shapes_to_test=shape_uris,
         target_class=target_class,
     )
