@@ -6,9 +6,9 @@ from flask import Blueprint, current_app, jsonify, request
 from flask_api import status
 from rdflib import Literal, URIRef
 from rdflib.term import Node
-from sqlalchemy.orm.exc import NoResultFound
 
 from buildingmotif.api.serializers.template import serialize
+from buildingmotif.database.errors import ModelNotFound, TemplateNotFound
 from buildingmotif.dataclasses import Model, Template
 from buildingmotif.ingresses import CSVIngress, TemplateIngress
 
@@ -42,10 +42,8 @@ def get_template(templates_id: int) -> flask.Response:
         template = current_app.building_motif.table_connection.get_db_template(
             templates_id
         )
-    except NoResultFound:
-        return {
-            "message": f"No template with id {templates_id}"
-        }, status.HTTP_404_NOT_FOUND
+    except TemplateNotFound:
+        return {"message": f"ID: {templates_id}"}, status.HTTP_404_NOT_FOUND
 
     return jsonify(serialize(template, include_parameters)), status.HTTP_200_OK
 
@@ -55,10 +53,8 @@ def evaluate_ingress(template_id: int) -> flask.Response:
     # get template
     try:
         template = Template.load(template_id)
-    except NoResultFound:
-        return {
-            "message": f"No template with id {template_id}"
-        }, status.HTTP_404_NOT_FOUND
+    except TemplateNotFound:
+        return {"message": f"ID: {template_id}"}, status.HTTP_404_NOT_FOUND
 
     # get model
     model_id = request.args.get("model_id")
@@ -68,13 +64,13 @@ def evaluate_ingress(template_id: int) -> flask.Response:
         }, status.HTTP_400_BAD_REQUEST
     try:
         model = Model.load(model_id)
-    except NoResultFound:
-        return {"message": f"No model with id {model_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {model_id}"}, status.HTTP_404_NOT_FOUND
 
     # get file
     raw_data = flask.request.get_data()
     if raw_data is None:
-        return {"message": "no file recieved."}, status.HTTP_404_NOT_FOUND
+        return {"message": "no file recieved."}, status.HTTP_400_NOT_FOUND
 
     # evaluate template
     try:
@@ -107,10 +103,8 @@ def evaluate_bindings(template_id: int) -> flask.Response:
     """
     try:
         template = Template.load(template_id)
-    except NoResultFound:
-        return {
-            "message": f"No template with id {template_id}"
-        }, status.HTTP_404_NOT_FOUND
+    except TemplateNotFound:
+        return {"message": f"ID: {template_id}"}, status.HTTP_404_NOT_FOUND
 
     if request.content_type != "application/json":
         return {
@@ -122,8 +116,8 @@ def evaluate_bindings(template_id: int) -> flask.Response:
         return {"message": "body must contain 'model_id'"}, status.HTTP_400_BAD_REQUEST
     try:
         model = Model.load(model_id)
-    except NoResultFound:
-        return {"message": f"No model with id {model_id}"}, status.HTTP_404_NOT_FOUND
+    except ModelNotFound:
+        return {"message": f"ID: {model_id}"}, status.HTTP_404_NOT_FOUND
 
     bindings = request.get_json().get("bindings")
     if bindings is None:
@@ -139,6 +133,23 @@ def evaluate_bindings(template_id: int) -> flask.Response:
         graph = graph_or_template
 
     return graph.serialize(format="ttl"), status.HTTP_200_OK
+
+
+@blueprint.route("/<template_id>/body", methods=(["GET"]))
+def get_template_body(template_id: int) -> flask.Response:
+    """Get template body.
+
+    :param template_id: template id
+    :type template_id: int
+    :return: template body
+    :rtype: flask.Response
+    """
+    try:
+        template: Template = Template.load(template_id)
+    except TemplateNotFound:
+        return {"message": f"ID: {template_id}"}, status.HTTP_404_NOT_FOUND
+
+    return template.body.serialize(format="ttl"), status.HTTP_200_OK
 
 
 def get_bindings(binding_dict) -> Dict[str, Node]:
