@@ -2,11 +2,10 @@ import logging
 import pathlib
 import tempfile
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Dict, List, Mapping, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 import pygit2
 import rdflib
-import sqlalchemy
 import yaml
 from pkg_resources import resource_exists, resource_filename
 from rdflib.exceptions import ParserError
@@ -14,78 +13,16 @@ from rdflib.plugins.parsers.notation3 import BadSyntax
 from rdflib.util import guess_format
 
 from buildingmotif import get_building_motif
+from buildingmotif.database.errors import LibraryNotFound
 from buildingmotif.database.tables import DBLibrary, DBTemplate
 from buildingmotif.dataclasses.shape_collection import ShapeCollection
 from buildingmotif.dataclasses.template import Template
 from buildingmotif.schemas import validate_libraries_yaml
 from buildingmotif.template_compilation import compile_template_spec
-from buildingmotif.utils import (
-    copy_graph,
-    get_ontology_files,
-    get_template_parts_from_shape,
-    shacl_inference,
-    skip_uri,
-)
+from buildingmotif.utils import get_ontology_files, shacl_inference
 
 if TYPE_CHECKING:
     from buildingmotif import BuildingMOTIF
-
-
-@dataclass
-class _template_dependency:
-    """Represents early-bound (template_id) or late-bound (template_name and
-    library) dependency of a template on another template.
-    """
-
-    template_name: str
-    bindings: Dict[str, Any]
-    library: str
-    template_id: Optional[int] = None
-
-    def __repr__(self):
-        return (
-            f"dep<name={self.template_name} bindings={self.bindings} "
-            f"library={self.library} id={self.template_id}>"
-        )
-
-    @classmethod
-    def from_dict(
-        cls, d: Dict[str, Any], dependent_library_name: str
-    ) -> "_template_dependency":
-        """Creates a py:class:`_template_dependency` from a dictionary.
-
-        :param d: dictionary
-        :type d: Dict[str, Any]
-        :param dependent_library_name: library name
-        :type dependent_library_name: str
-        :return: the _template_dependency from the dict
-        :rtype: _template_dependency
-        """
-        template_name = d["template"]
-        bindings = d.get("args", {})
-        library = d.get("library", dependent_library_name)
-        template_id = d.get("template_id")
-        return cls(template_name, bindings, library, template_id)
-
-    def to_template(self, id_lookup: Dict[str, int]) -> Template:
-        """Resolve this dependency to a template.
-
-        :param id_lookup: a local cache of {name: id} for uncommitted templates
-        :type id_lookup: Dict[str, int]
-        :return: the template instance this dependency points to
-        :rtype: Template
-        """
-        # direct lookup if id is provided
-        if self.template_id is not None:
-            return Template.load(self.template_id)
-        # if id is not provided, look at our local 'cache' of to-be-committed
-        # templates for the id (id_lookup)
-        if self.template_name in id_lookup:
-            return Template.load(id_lookup[self.template_name])
-        # if not in the local cache, then search the database for the template
-        # within the given library
-        library = Library.load(name=self.library)
-        return library.get_template_by_name(self.template_name)
 
 
 @dataclass
@@ -116,7 +53,7 @@ class Library:
                 logging.warning(
                     f'Library {name} already exists in database. To ovewrite load library with "overwrite=True"'  # noqa
                 )
-        except sqlalchemy.exc.NoResultFound:
+        except LibraryNotFound:
             db_library = bm.table_connection.create_db_library(name)
 
         return cls(_id=db_library.id, _name=db_library.name, _bm=bm)
@@ -282,54 +219,17 @@ class Library:
 
         lib = cls.create(ontology_name, overwrite=overwrite)
 
-        if infer_templates:
-            # infer shapes from any class/nodeshape candidates in the graph
-            lib._infer_templates_from_graph(ontology)
-
         # load the ontology graph as a shape_collection
         shape_col_id = lib.get_shape_collection().id
         assert shape_col_id is not None  # should always pass
         shape_col = ShapeCollection.load(shape_col_id)
         shape_col.add_graph(ontology)
 
+        if infer_templates:
+            # infer shapes from any class/nodeshape candidates in the graph
+            shape_col.infer_templates(lib)
+
         return lib
-
-    def _infer_templates_from_graph(self, graph: rdflib.Graph):
-        """Infer templates from a graph (by interpreting shapes) and add them to this library.
-
-        :param graph: graph to infer templates from
-        :type graph: rdflib.Graph
-        """
-        # add all imports to the same graph so we can resolve everything
-        imports_closure = copy_graph(graph)
-        # import dependencies into 'graph'
-        # get all imports from the graph
-        for dependency in graph.objects(predicate=rdflib.OWL.imports):
-            # attempt to load from BuildingMOTIF
-            try:
-                lib = Library.load(name=str(dependency))
-                imports_closure += lib.get_shape_collection().graph
-            except Exception as e:  # TODO: replace with a more specific exception
-                logging.warning(
-                    f"An ontology could not resolve a dependency on {dependency} ({e}). Check this is loaded into BuildingMOTIF"
-                )
-                continue
-        class_candidates = set(graph.subjects(rdflib.RDF.type, rdflib.OWL.Class))
-        shape_candidates = set(graph.subjects(rdflib.RDF.type, rdflib.SH.NodeShape))
-        candidates = class_candidates.intersection(shape_candidates)
-        template_id_lookup: Dict[str, int] = {}
-        dependency_cache: Dict[int, List[Dict[Any, Any]]] = {}
-        for candidate in candidates:
-            assert isinstance(candidate, rdflib.URIRef)
-            # TODO: mincount 0 (or unspecified) should be optional args on the generated template
-            partial_body, deps = get_template_parts_from_shape(
-                candidate, imports_closure
-            )
-            templ = self.create_template(str(candidate), partial_body)
-            dependency_cache[templ.id] = deps
-            template_id_lookup[str(candidate)] = templ.id
-
-        self._resolve_template_dependencies(template_id_lookup, dependency_cache)
 
     def _load_shapes_from_directory(
         self,
@@ -364,7 +264,7 @@ class Library:
             )
         # infer shapes from any class/nodeshape candidates in the graph
         if infer_templates:
-            self._infer_templates_from_graph(shape_col.graph)
+            shape_col.infer_templates(self)
 
     @classmethod
     def _load_from_directory(
@@ -403,18 +303,12 @@ class Library:
 
         lib = cls.create(directory.name, overwrite=overwrite)
 
-        # setup caches for reading templates
-        template_id_lookup: Dict[str, int] = {}
-        dependency_cache: Dict[int, List[_template_dependency]] = {}
-
         # read all .yml files
         for file in directory.rglob("*.yml"):
             # if .ipynb_checkpoints, skip; these are cached files that Jupyter creates
             if ".ipynb_checkpoints" in file.parts:
                 continue
-            lib._read_yml_file(file, template_id_lookup, dependency_cache)
-        # now that we have all the templates, we can populate the dependencies
-        lib._resolve_template_dependencies(template_id_lookup, dependency_cache)
+            lib._read_yml_file(file)
         # load shape collections from all ontology files in the directory
         lib._load_shapes_from_directory(directory)
 
@@ -445,89 +339,10 @@ class Library:
         try:
             bm.table_connection.get_db_library_by_name(library_name)
             return True
-        except sqlalchemy.exc.NoResultFound:
+        except LibraryNotFound:
             return False
 
-    def _resolve_dependency(
-        self,
-        template: Template,
-        dep: Union[_template_dependency, dict],
-        template_id_lookup: Dict[str, int],
-    ):
-        """Resolve a dependency to a template.
-
-        :param template: template to resolve dependency for
-        :type template: Template
-        :param dep: dependency
-        :type dep: Union[_template_dependency, dict]
-        :param template_id_lookup: a local cache of {name: id} for uncommitted templates
-        :type template_id_lookup: Dict[str, int]
-        :return: the template instance this dependency points to
-        :rtype: Template
-        """
-        # if dep is a _template_dependency, turn it into a template
-        if isinstance(dep, _template_dependency):
-            dependee = dep.to_template(template_id_lookup)
-            template.add_dependency(dependee, dep.bindings)
-            return
-
-        # now, we know that dep is a dict
-
-        # if dependency names a library explicitly, load that library and get the template by name
-        if "library" in dep:
-            dependee = Library.load(name=dep["library"]).get_template_by_name(
-                dep["template"]
-            )
-            template.add_dependency(dependee, dep["args"])
-            return
-        # if no library is provided, try to resolve the dependency from this library
-        if dep["template"] in template_id_lookup:
-            dependee = Template.load(template_id_lookup[dep["template"]])
-            template.add_dependency(dependee, dep["args"])
-            return
-        # check documentation for skip_uri for what URIs get skipped
-        if skip_uri(dep["template"]):
-            return
-
-        # if the dependency is not in the local cache, then search through this library's imports
-        # for the template
-        for imp in self.graph_imports:
-            try:
-                library = Library.load(name=str(imp))
-                dependee = library.get_template_by_name(dep["template"])
-                template.add_dependency(dependee, dep["args"])
-                return
-            except Exception as e:
-                logging.debug(
-                    f"Could not find dependee {dep['template']} in library {imp}: {e}"
-                )
-        logging.warning(
-            f"Warning: could not find dependee {dep['template']} in libraries {self.graph_imports}"
-        )
-
-    def _resolve_template_dependencies(
-        self,
-        template_id_lookup: Dict[str, int],
-        dependency_cache: Mapping[int, Union[List[_template_dependency], List[dict]]],
-    ):
-        """Resolve all dependencies for all templates in this library"""
-        # two phases here: first, add all of the templates and their dependencies
-        # to the database but *don't* check that the dependencies are valid yet
-        for template in self.get_templates():
-            if template.id not in dependency_cache:
-                continue
-            for dep in dependency_cache[template.id]:
-                self._resolve_dependency(template, dep, template_id_lookup)
-        # check that all dependencies are valid (use parameters that exist, etc)
-        for template in self.get_templates():
-            template.check_dependencies()
-
-    def _read_yml_file(
-        self,
-        file: pathlib.Path,
-        template_id_lookup: Dict[str, int],
-        dependency_cache: Dict[int, List[_template_dependency]],
-    ):
+    def _read_yml_file(self, file: pathlib.Path):
         """Read a YML file into this library. Utility function for `_load_from_directory`."""
         contents = yaml.load(open(file, "r"), Loader=yaml.FullLoader)
         for templ_name, templ_spec in contents.items():
@@ -536,20 +351,14 @@ class Library:
             # input name of template
             templ_spec.update({"name": templ_name})
             # remove dependencies so we can resolve them to their IDs later
-            deps = [
-                _template_dependency.from_dict(d, self.name)
-                for d in templ_spec.pop("dependencies", [])
-            ]
             templ_spec["optional_args"] = templ_spec.pop("optional", [])
             try:
-                templ = self.create_template(**templ_spec)
+                self.create_template(**templ_spec)
             except Exception as e:
                 logging.error(
                     f"Error creating template {templ_name} from file {file}: {e}"
                 )
                 raise e
-            dependency_cache[templ.id] = deps
-            template_id_lookup[templ.name] = templ.id
 
     @property
     def id(self) -> Optional[int]:
@@ -585,6 +394,7 @@ class Library:
         name: str,
         body: Optional[rdflib.Graph] = None,
         optional_args: Optional[List[str]] = None,
+        dependencies: Optional[List] = None,
     ) -> Template:
         """Create template in this library.
 
@@ -608,6 +418,22 @@ class Library:
         self._bm.table_connection.update_db_template_optional_args(
             db_template.id, optional_args
         )
+
+        if dependencies is not None:
+            for dependency in dependencies:
+                dependency_template = dependency["template"]
+                dependency_library = None
+                if "library" in dependency:
+                    dependency_library = dependency["library"]
+                else:
+                    dependency_library = self.name
+                dependency_args = dependency["args"]
+                self._bm.table_connection.add_template_dependency_preliminary(
+                    db_template.id,
+                    dependency_library,
+                    dependency_template,
+                    dependency_args,
+                )
 
         return Template(
             _id=db_template.id,
