@@ -12,8 +12,8 @@ from rdflib import BNode, Graph, Literal, URIRef
 from rdflib.compare import _TripleCanonicalizer
 from rdflib.paths import ZeroOrOne
 from rdflib.term import Node
-from sqlalchemy.exc import NoResultFound
 
+from buildingmotif.database.errors import TemplateNotFound
 from buildingmotif.namespaces import OWL, PARAM, RDF, SH, XSD, bind_prefixes
 
 if TYPE_CHECKING:
@@ -63,7 +63,7 @@ def _guarantee_unique_template_name(library: "Library", name: str) -> str:
         while library.get_template_by_name(name):
             name = f"{original_name}_{idx}"
             idx += 1
-    except NoResultFound:
+    except TemplateNotFound:
         # this means that the template does not exist and we can use the original name
         pass
     return name
@@ -214,7 +214,9 @@ def get_ontology_files(directory: Path, recursive: bool = True) -> List[Path]:
 
 
 def get_template_parts_from_shape(
-    shape_name: URIRef, shape_graph: Graph
+    shape_name: URIRef,
+    shape_graph: Graph,
+    depedency_graphs: Dict[str, Graph] = {},
 ) -> Tuple[Graph, List[Dict]]:
     """Turn a SHACL shape into a template. The following attributes of
     NodeShapes will be incorporated into the resulting template:
@@ -227,6 +229,8 @@ def get_template_parts_from_shape(
     :type shape_name: URIRef
     :param shape_graph: shape graph
     :type shape_graph: Graph
+    :param depedency_graphs: colleciton of graphs and which depdency library they came from
+    :type depedency_graphs: dict[str, Graph]
     :raises Exception: if more than one object type detected on shape
     :raises Exception: if more than one min count detected on shape
     :return: template parts
@@ -285,7 +289,23 @@ def get_template_parts_from_shape(
             otype_is_nodeshape = (otype, RDF.type, SH.NodeShape) in shape_graph
 
             if (otype_as_class and otype_is_nodeshape) or otype_as_node:
-                deps.append({"template": str(otype), "args": {"name": param}})
+                if not isinstance(otype, URIRef):
+                    continue
+                library = None
+                for library_name, graph in depedency_graphs.items():
+                    if (otype, RDF.type, SH.NodeShape) in graph:
+                        library = library_name
+                        break
+                if library is None:
+                    deps.append({"template": str(otype), "args": {"name": param}})
+                else:
+                    deps.append(
+                        {
+                            "template": str(otype),
+                            "library": library,
+                            "args": {"name": param},
+                        }
+                    )
                 body.add((param, RDF.type, otype))
 
         pvalue = shape_graph.value(pshape, SH["hasValue"])
@@ -309,6 +329,9 @@ def get_template_parts_from_shape(
     for node in nodes:
         # if node is already in deps, skip it
         if any(str(node) == dep["template"] for dep in deps):
+            continue
+        # skip non-URIRef nodes
+        if not isinstance(node, URIRef):
             continue
         deps.append(
             {"template": str(node), "args": {"name": "name"}}
@@ -338,7 +361,9 @@ def _prep_shape_graph() -> Graph:
     return shape
 
 
-def _index_properties(templ: "Template") -> _TemplateIndex:
+def _index_properties(
+    templ: "Template", error_on_missing_dependency: bool = True
+) -> _TemplateIndex:
     templ_graph = templ.evaluate(
         {p: PARAM[p] for p in templ.parameters}, {"mark": PARAM}
     )
@@ -365,7 +390,9 @@ def _index_properties(templ: "Template") -> _TemplateIndex:
         # maybe_param = str(o).removeprefix(PARAM) Python >=3.9
         maybe_param = str(o)[len(PARAM) :]
         if maybe_param in templ.dependency_parameters:
-            dep = templ.dependency_for_parameter(maybe_param)
+            dep = templ.dependency_for_parameter(
+                maybe_param, error_on_missing_dependency
+            )
             if dep is not None:
                 prop_shapes[p].append(URIRef(dep._name))
         elif o in param_types:
@@ -538,6 +565,25 @@ def _inline_sh_and(sg: Graph):
             sg.add((parent, p, o))
 
 
+def _inline_sh_qualified_value_shape(sg: Graph):
+    """
+    This detects the use of 'sh:qualifiedValueShape' on SHACL PropertyShapes and inlines
+    all of the included shapes
+    """
+    q = """
+    PREFIX sh: <http://www.w3.org/ns/shacl#>
+    SELECT ?parent ?child WHERE {
+        ?parent a sh:PropertyShape ;
+                sh:qualifiedValueShape ?child .
+        }"""
+    for row in sg.query(q):
+        parent, child = row  # type: ignore
+        sg.remove((parent, SH["qualifiedValueShape"], child))
+        pobjs = sg.predicate_objects(child)
+        for (p, o) in pobjs:
+            sg.add((parent, p, o))
+
+
 def rewrite_shape_graph(g: Graph) -> Graph:
     """
     Rewrites the input graph to make the resulting validation report more useful.
@@ -555,6 +601,7 @@ def rewrite_shape_graph(g: Graph) -> Graph:
         _inline_sh_and(sg)
         # make sure to handle sh:node *after* sh:and
         _inline_sh_node(sg)
+        _inline_sh_qualified_value_shape(sg)
     return sg
 
 
