@@ -28,14 +28,14 @@ def setup_building_motif_brick() -> Tuple[BuildingMOTIF, Library]:
     )
     dependency_graphs = [
         "libraries/brick/imports/ref-schema.ttl",
-        "libraries/qudt/VOCAB_QUDT-QUANTITY-KINDS-ALL-v2.1.ttl",
-        "libraries/qudt/VOCAB_QUDT-DIMENSION-VECTORS-v2.1.ttl",
-        "libraries/qudt/VOCAB_QUDT-UNITS-ALL-v2.1.ttl",
-        "libraries/qudt/SCHEMA-FACADE_QUDT-v2.1.ttl",
-        "libraries/qudt/SCHEMA_QUDT_NoOWL-v2.1.ttl",
-        "libraries/qudt/VOCAB_QUDT-PREFIXES-v2.1.ttl",
-        "libraries/qudt/SHACL-SCHEMA-SUPPLEMENT_QUDT-v2.1.ttl",
-        "libraries/qudt/VOCAB_QUDT-SYSTEM-OF-UNITS-ALL-v2.1.ttl",
+        "libraries/qudt/VOCAB_QUDT-QUANTITY-KINDS-ALL.ttl",
+        "libraries/qudt/VOCAB_QUDT-DIMENSION-VECTORS.ttl",
+        "libraries/qudt/VOCAB_QUDT-UNITS-ALL.ttl",
+        "libraries/qudt/SCHEMA-FACADE_QUDT.ttl",
+        "libraries/qudt/SCHEMA_QUDT_NoOWL.ttl",
+        "libraries/qudt/VOCAB_QUDT-PREFIXES.ttl",
+        "libraries/qudt/SHACL-SCHEMA-SUPPLEMENT_QUDT.ttl",
+        "libraries/qudt/VOCAB_QUDT-SYSTEM-OF-UNITS-ALL.ttl",
         "libraries/brick/imports/rec.ttl",
         "libraries/brick/imports/recimports.ttl",
         "libraries/brick/imports/brickpatches.ttl",
@@ -77,7 +77,7 @@ def pytest_generate_tests(metafunc):
         ids = []
         for library_name in libraries:
             library = Library.load(directory=library_name, run_shacl_inference=False)
-            templates = library.get_templates()
+            templates = sorted(library.get_templates(), key=lambda t: t.name)
             params.extend([(bm, brick, library, template) for template in templates])
             ids.extend([f"{library.name}-{template.name}" for template in templates])
         metafunc.parametrize("bm,brick,library,template", params, ids=ids)
