@@ -1,12 +1,12 @@
 import uuid
 
 import pytest
-from sqlalchemy.exc import NoResultFound
 
+from buildingmotif.database.errors import ShapeCollectionNotFound
 from buildingmotif.database.tables import DBShapeCollection
 
 
-def test_create_db_shape_collection(monkeypatch, table_connection):
+def test_create_db_shape_collection(monkeypatch, bm):
     mocked_uuid = uuid.uuid4()
 
     def mockreturn():
@@ -14,16 +14,16 @@ def test_create_db_shape_collection(monkeypatch, table_connection):
 
     monkeypatch.setattr(uuid, "uuid4", mockreturn)
 
-    db_shape_collection = table_connection.create_db_shape_collection()
+    db_shape_collection = bm.table_connection.create_db_shape_collection()
 
     assert db_shape_collection.graph_id == str(mocked_uuid)
 
 
-def test_get_db_shape_collections(table_connection):
-    shape_collection1 = table_connection.create_db_shape_collection()
-    shape_collection2 = table_connection.create_db_shape_collection()
+def test_get_db_shape_collections(bm):
+    shape_collection1 = bm.table_connection.create_db_shape_collection()
+    shape_collection2 = bm.table_connection.create_db_shape_collection()
 
-    db_shape_collections = table_connection.get_all_db_shape_collections()
+    db_shape_collections = bm.table_connection.get_all_db_shape_collections()
 
     assert len(db_shape_collections) == 2
     assert all(type(m) == DBShapeCollection for m in db_shape_collections)
@@ -47,7 +47,7 @@ def test_get_db_shape_collection(table_connection, monkeypatch):
 
 
 def test_get_db_shape_collection_does_not_exist(table_connection):
-    with pytest.raises(NoResultFound):
+    with pytest.raises(ShapeCollectionNotFound):
         table_connection.get_db_shape_collection("I don't exist")
 
 
@@ -55,10 +55,10 @@ def test_delete_db_shape_collection(table_connection):
     db_shape_collection = table_connection.create_db_shape_collection()
     table_connection.delete_db_shape_collection(db_shape_collection.id)
 
-    with pytest.raises(NoResultFound):
+    with pytest.raises(ShapeCollectionNotFound):
         table_connection.get_db_shape_collection(db_shape_collection.id)
 
 
-def tests_delete_db_shape_collection_does_does_exist(table_connection):
-    with pytest.raises(NoResultFound):
+def tests_delete_db_shape_collection_does_not_exist(table_connection):
+    with pytest.raises(ShapeCollectionNotFound):
         table_connection.delete_db_shape_collection("does not exist")
