@@ -5,6 +5,7 @@ import rdflib
 from sqlalchemy.exc import IntegrityError, NoResultFound
 
 from buildingmotif import BuildingMOTIF
+from buildingmotif.database.errors import LibraryNotFound, TemplateNotFound
 from buildingmotif.database.tables import DBTemplate
 
 
@@ -77,7 +78,7 @@ def test_create_db_template_bad_library(bm: BuildingMOTIF, monkeypatch):
 
     monkeypatch.setattr(uuid, "uuid4", mockreturn)
 
-    with pytest.raises(NoResultFound):
+    with pytest.raises(LibraryNotFound):
         bm.table_connection.create_db_template(
             name="my_db_template",
             library_id=-999,  # id does not exist
@@ -139,7 +140,7 @@ def test_get_db_template(bm: BuildingMOTIF, monkeypatch):
 
 
 def test_get_db_template_does_not_exist(bm: BuildingMOTIF):
-    with pytest.raises(NoResultFound):
+    with pytest.raises(TemplateNotFound):
         bm.table_connection.get_db_template(-999)
 
 
@@ -176,7 +177,7 @@ def test_update_db_template_name_bad_name(bm: BuildingMOTIF):
 
 
 def test_update_db_template_name_does_not_exist(bm: BuildingMOTIF):
-    with pytest.raises(NoResultFound):
+    with pytest.raises(TemplateNotFound):
         bm.table_connection.update_db_template_name(-999, "new_name")
 
 
@@ -188,12 +189,12 @@ def test_delete_db_template(bm: BuildingMOTIF):
 
     bm.table_connection.delete_db_template(db_template.id)
 
-    with pytest.raises(NoResultFound):
-        bm.table_connection.get_db_model(db_template.id)
+    with pytest.raises(TemplateNotFound):
+        bm.table_connection.get_db_template(db_template.id)
 
 
 def tests_delete_db_template_does_does_exist(bm: BuildingMOTIF):
-    with pytest.raises(NoResultFound):
+    with pytest.raises(TemplateNotFound):
         bm.table_connection.delete_db_template(-999)
 
 
@@ -201,21 +202,24 @@ def test_add_template_dependency(bm: BuildingMOTIF):
     (
         _,
         dependant_template,
-        dependee_template,
+        dependency_template,
     ) = create_dependency_test_fixtures(bm)
 
     bm.table_connection.add_template_dependency_preliminary(
-        dependant_template.id, dependee_template.id, {"name": "ding", "h2": "dong"}
+        dependant_template.id,
+        dependency_template.library.name,
+        dependency_template.name,
+        {"name": "ding", "h2": "dong"},
     )
     bm.table_connection.check_all_template_dependencies()
 
-    assert dependant_template.dependencies == [dependee_template]
-    assert dependee_template.dependants == [dependant_template]
+    assert dependant_template.dependencies[0].dependency_template == dependency_template
+
     res = bm.table_connection.get_db_template_dependencies(dependant_template.id)
     assert len(res) == 1
     dep_assoc = res[0]
-    assert dep_assoc.dependant_id == dependant_template.id
-    assert dep_assoc.dependee_id == dependee_template.id
+    assert dep_assoc.template_id == dependant_template.id
+    assert dep_assoc.dependency_template == dependency_template
     assert dep_assoc.args == {"name": "ding", "h2": "dong"}
 
 
@@ -228,7 +232,10 @@ def test_add_template_dependency_bad_args(bm: BuildingMOTIF):
 
     with pytest.raises(ValueError):
         bm.table_connection.add_template_dependency_preliminary(
-            dependant_template.id, dependee_template.id, {"bad": "ding"}
+            dependant_template.id,
+            dependee_template.library.name,
+            dependee_template.name,
+            {"bad": "ding"},
         )
         bm.table_connection.check_all_template_dependencies()
 
@@ -241,12 +248,18 @@ def test_add_template_dependency_already_exist(bm: BuildingMOTIF):
     ) = create_dependency_test_fixtures(bm)
 
     bm.table_connection.add_template_dependency_preliminary(
-        dependant_template.id, dependee_template.id, {"name": "ding", "h2": "dong"}
+        dependant_template.id,
+        dependee_template.library.name,
+        dependee_template.name,
+        {"name": "ding", "h2": "dong"},
     )
 
     with pytest.raises(IntegrityError):
         bm.table_connection.add_template_dependency_preliminary(
-            dependant_template.id, dependee_template.id, {"name": "ding", "h2": "dong"}
+            dependant_template.id,
+            dependee_template.library.name,
+            dependee_template.name,
+            {"name": "ding", "h2": "dong"},
         )
         bm.table_connection.check_all_template_dependencies()
 
@@ -261,15 +274,18 @@ def test_get_dependencies(bm: BuildingMOTIF):
     ) = create_dependency_test_fixtures(bm)
 
     bm.table_connection.add_template_dependency_preliminary(
-        dependant_template.id, dependee_template.id, {"name": "ding", "h2": "dong"}
+        dependant_template.id,
+        dependee_template.library.name,
+        dependee_template.name,
+        {"name": "ding", "h2": "dong"},
     )
     bm.table_connection.check_all_template_dependencies()
 
     res = bm.table_connection.get_db_template_dependencies(dependant_template.id)
     assert len(res) == 1
     dep_assoc = res[0]
-    assert dep_assoc.dependant_id == dependant_template.id
-    assert dep_assoc.dependee_id == dependee_template.id
+    assert dep_assoc.template_id == dependant_template.id
+    assert dep_assoc.dependency_template == dependee_template
     assert dep_assoc.args == {"name": "ding", "h2": "dong"}
 
 
@@ -281,15 +297,18 @@ def test_remove_dependencies(bm: BuildingMOTIF):
     ) = create_dependency_test_fixtures(bm)
 
     bm.table_connection.add_template_dependency_preliminary(
-        dependant_template.id, dependee_template.id, {"name": "ding", "h2": "dong"}
+        dependant_template.id,
+        dependee_template.library.name,
+        dependee_template.name,
+        {"name": "ding", "h2": "dong"},
     )
     bm.table_connection.check_all_template_dependencies()
 
     res = bm.table_connection.get_db_template_dependencies(dependant_template.id)
     assert len(res) == 1
     dep_assoc = res[0]
-    assert dep_assoc.dependant_id == dependant_template.id
-    assert dep_assoc.dependee_id == dependee_template.id
+    assert dep_assoc.template_id == dependant_template.id
+    assert dep_assoc.dependency_template == dependee_template
     assert dep_assoc.args == {"name": "ding", "h2": "dong"}
 
     bm.table_connection.delete_template_dependency(
